requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
bertopic==0.15.0
umap-learn==0.5.4
nltk==3.8.1
gradio==3.50.2
matplotlib==3.7.2
seaborn==0.12.2
jsonschema==4.19.0
numba==0.58.1